process_wrap: update after libuv api change
The `exit_code` argument for the `ProcessWrap::OnExit` callback changed
from int to int64_t.

+++ src/process_wrap.cc
@@ -265,7 +265,9 @@ class ProcessWrap : public HandleWrap {
     args.GetReturnValue().Set(err);
   }
 
-  static void OnExit(uv_process_t* handle, int exit_status, int term_signal) {
+  static void OnExit(uv_process_t* handle,
+                     int64_t exit_status,
+                     int term_signal) {
     HandleScope scope(node_isolate);
 
     ProcessWrap* wrap = static_cast<ProcessWrap*>(handle->data);
